+++ plom/finish/coded_return.py
@@ -17,7 +17,7 @@ import shutil
 from pathlib import Path
 
 from plom import SpecParser
-from plom.rules import isValidStudentNumber
+from plom.rules import isValidStudentNumber, StudentIDLength
 from plom.finish import CSVFilename
 from .return_tools import csv_add_return_codes
 
@@ -45,7 +45,14 @@ def do_renaming(fromdir, todir, sns):
     return numfiles
 
 
-def main():
+def main(use_hex, digits):
+    """Make the secret codes and the return-code webpage.
+
+    args:
+        use_hex (bool): use random hex digits, otherwise an integer
+            without leading zeros.
+        digits (int): length of secret code.
+    """
     spec = SpecParser().spec
     shortname = spec["name"]
     longname = spec["longName"]
@@ -77,7 +84,9 @@ def main():
     os.makedirs(codedReturnDir)
 
     print("Generating return codes spreadsheet...")
-    sns = csv_add_return_codes(CSVFilename, "return_codes.csv", "StudentID")
+    sns = csv_add_return_codes(
+        CSVFilename, "return_codes.csv", "StudentID", use_hex, digits
+    )
     print('The return codes are in "return_codes.csv"')
 
     numfiles = do_renaming(fromdir, codedReturnDir, sns)
@@ -92,6 +101,8 @@ def main():
 
     html = html.replace("__COURSENAME__", longname)
     html = html.replace("__TESTNAME__", shortname)
+    html = html.replace("__CODE_LENGTH__", str(digits))
+    html = html.replace("__SID_LENGTH__", str(StudentIDLength))
 
     with open(codedReturnDir / "index.html", "w") as htmlfile:
         htmlfile.write(html)
@@ -101,7 +112,3 @@ def main():
     print('  * Privately communicate info from "return_codes.csv"')
     print("      - E.g., see `contrib/plom-return_codes_to_canvas_csv.py`")
     print("  * Read docs about the security implications of all this.")
-
-
-if __name__ == "__main__":
-    main()