+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* New command line arguments for `plom-finish` for digital return.
+* Canvas-related return code handing reduced from 12 digits to 9 by default.
 
 ### Fixed